+++ src/test/kotlin/tornadofx/NodeTests.kt
@@ -1,7 +1,5 @@
 package tornadofx
 
-import javafx.application.Platform
-import javafx.scene.Parent
 import javafx.scene.control.Label
 import javafx.scene.control.TableColumn
 import javafx.scene.layout.HBox